requirements.txt
lxml==4.4.1
Pillow==6.1.0
requests==2.22.0
tqdm==4.36.1
xmltodict==0.12.0
wikipedia==1.4.0
Django==2.2.7
numpy==1.15.1
pytest==3.7.4
rope==0.11.0
autopep8==1.4
yapf==0.23.0
gunicorn==19.7.1
eventlet==0.21.0
gevent==1.4.0
pandas==0.25.2
django-extensions==2.2.5